refactor(market-list): 로직과 UI 분리 및 FSD 아키텍처 개선
- FSD 아키텍처 위반 수정
  - price-change.ts를 features/upbit-chart에서 entities/upbit으로 이동
  - feature-to-feature 의존성 제거
  - calculatePriceChange에서 하드코딩된 색상 제거, CHANGE_TYPE_COLORS 사용

- 공통 유틸리티 통합
  - formatTradePrice를 entities/upbit/lib/format.ts로 이동
  - 중복 getChangeColor 함수 제거
  - COLORS 상수를 CHANGE_TYPE_COLORS로 통일

- market-list hooks 생성 (로직/UI 분리)
  - useMarketListData: 데이터 fetching & WebSocket 관리
  - useMarketListSort: 정렬 로직
  - useMarketListFavorites: 즐겨찾기 관리
  - useMarketListHighlights: 가격 하이라이트 애니메이션

- MarketListTable 리팩토링
  - 인라인 로직을 hooks로 대체 (194줄 감소)
  - useState/useEffect/useMemo 제거하고 커스텀 hooks 사용
  - 코드 가독성 및 재사용성 향상

- Storybook 개선
  - 5개의 추가 스토리 (가격순, 변동률순, 이름순, 클릭 핸들러)
  - 레거시 경고 메시지 추가

- Import 경로 정리
  - candlestick-chart.tsx의 import 경로 수정
  - entities/upbit에 formatTradePrice, calculatePriceChange export 추가
  - features/upbit-chart에서 price-change export 제거
