+++ src/features/market-list/ui/market-list-table.stories.tsx
@@ -34,13 +34,65 @@ export default meta;
 type Story = StoryObj<typeof meta>;
 
 /**
- * 기본 마켓 리스트 테이블
+ * 기본 마켓 리스트 테이블 (기존 컴포넌트)
  * - REST API로 KRW 마켓 목록과 초기 현재가 조회
  * - WebSocket으로 실시간 가격 업데이트
  * - 거래대금 순으로 정렬 (기본)
+ *
+ * ⚠️ 레거시 컴포넌트: 하위 호환성을 위해 유지됩니다.
+ * 새로운 프로젝트는 MarketList (Compound Component)를 사용하세요.
  */
 export const Default: Story = {
   args: {
     initialSortBy: 'acc_trade_price_24h',
+    initialSortOrder: 'desc',
+  },
+};
+
+/**
+ * 가격 순 정렬
+ * - 현재가 기준 내림차순 정렬
+ */
+export const SortByPrice: Story = {
+  args: {
+    initialSortBy: 'trade_price',
+    initialSortOrder: 'desc',
+  },
+};
+
+/**
+ * 변동률 순 정렬
+ * - 전일대비 변동률 기준 내림차순 정렬
+ */
+export const SortByChangeRate: Story = {
+  args: {
+    initialSortBy: 'change_rate',
+    initialSortOrder: 'desc',
+  },
+};
+
+/**
+ * 코인명 순 정렬 (가나다순)
+ * - 한글명 기준 오름차순 정렬
+ */
+export const SortByName: Story = {
+  args: {
+    initialSortBy: 'korean_name',
+    initialSortOrder: 'asc',
+  },
+};
+
+/**
+ * Row 클릭 핸들러
+ * - onRowClick 콜백으로 클릭 이벤트 처리
+ */
+export const WithRowClick: Story = {
+  args: {
+    initialSortBy: 'acc_trade_price_24h',
+    initialSortOrder: 'desc',
+    onRowClick: (market: string) => {
+      console.log('Clicked:', market);
+      alert(`선택된 마켓: ${market}`);
+    },
   },
 };
